# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Task API",
    "version": "1.0.0",
    "description": "CRUD API built using Express.js"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
    "/tasks": {
      "get": {
        "summary": "Get all tasks",
        "responses": {
          "200": {
            "description": "List of tasks"
          }
        }
      },
      "post": {
        "summary": "Create a new task",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  }
                }
              },
              "example": {
                "title": "Buy milk"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task created"
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/tasks/{id}": {
      "put": {
        "summary": "Update a task",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "example": {
                "title": "Buy Bread",
                "done": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task updated"
          },
          "400": {
            "description": "Invalid request"
          },
          "404": {
            "description": "Task not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a task",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Task deleted"
          },
          "404": {
            "description": "Task not found"
          }
        }
      }
    }
  }
}